pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    { name = "abderrahimarf", email = "arfaouiabderahime2002@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.11,<3.12"

dependencies = [
    "sentence-transformers >=5.0.0,<6.0.0",
    "pymupdf >=1.26.3,<2.0.0",
    "python-docx >=1.2.0,<2.0.0",
    "langgraph >=0.6.2,<0.7.0",
    "langchain >=0.3.27,<0.4.0",
    "langchain-community >=0.3.27,<0.4.0",
    "openai >=1.98.0,<2.0.0",
    "langchain-openai >=0.3.28,<0.4.0",
    "weasyprint >=66.0,<67.0",
    "xhtml2pdf >=0.2.17,<0.3.0",
    "flask >=3.1.1,<4.0.0",
    "flask-cors >=6.0.1,<7.0.0",
    "flask-socketio[client] >=5.5.1,<6.0.0",
    "eventlet >=0.40.2,<0.41.0",
    "fastapi >=0.116.1,<0.117.0",
    "uvicorn[standard] >=0.35.0,<0.36.0",
    "python-socketio >=5.13.0,<6.0.0",
    "cerebras-cloud-sdk >=1.35.0,<2.0.0",
    "weaviate-client >=4.16.5,<5.0.0",
    "duckduckgo-search (>=8.1.1,<9.0.0)",
    "docling (>=2.43.0,<3.0.0)",
    "supabase (==2.5.1)",
    "python-dotenv (==1.0.1)",
    "itsdangerous (==2.2.0)",
    "pymongo (>=4.6)",
    "certifi (>=2025.8.3,<2026.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "datasets (>=4.0.0,<5.0.0)",
    "transformers (>=4.55.2,<5.0.0)",
    "accelerate (>=1.10.0,<2.0.0)",
    "scikit-learn (>=1.7.1,<2.0.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "waitress (>=3.0.2,<4.0.0)",
    "langchain-google-genai (<2.1.9)",
    "dotenv (>=0.9.9,<0.10.0)",
    "numpy (<2.0)",
    "lightgbm (>=4.6.0,<5.0.0)",
]

[tool.poetry]
packages = [{ include = "backend", from = "src" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
